Finalize challenge project
