+++ apps/openchallenges/app/src/app/app-routing.module.ts
@@ -68,7 +68,7 @@ export const routes: Routes = [
     path: 'challenge/:challengeId',
     loadChildren: () =>
       import('@sagebionetworks/openchallenges/challenge').then(
-        (m) => m.ChallengeModule
+        (routes) => routes.routes
       ),
   },
   // {
