--- libs/openchallenges/challenge/src/lib/challenge-routing.module.ts
@@ -1,17 +1,0 @@
-import { NgModule } from '@angular/core';
-import { RouterModule, Routes } from '@angular/router';
-import { ChallengeComponent } from './challenge.component';
-
-export const routes: Routes = [
-  {
-    path: '',
-    component: ChallengeComponent,
-    children: [{ path: '**', component: ChallengeComponent }],
-  },
-];
-
-@NgModule({
-  imports: [RouterModule.forChild(routes)],
-  exports: [RouterModule],
-})
-export class ChallengeRoutingModule {}
