+++ libs/openchallenges/challenge/src/lib/challenge.component.ts
@@ -1,5 +1,10 @@
 import { Component, OnInit } from '@angular/core';
-import { ActivatedRoute, ParamMap, Router } from '@angular/router';
+import {
+  ActivatedRoute,
+  ParamMap,
+  Router,
+  RouterModule,
+} from '@angular/router';
 import {
   Challenge,
   ChallengeService,
@@ -15,15 +20,43 @@ import {
 } from 'rxjs';
 import { Tab } from './tab.model';
 import { CHALLENGE_TABS } from './challenge-tabs';
-import { Avatar } from '@sagebionetworks/openchallenges/ui';
+import {
+  Avatar,
+  AvatarComponent,
+  FooterComponent,
+  UiModule,
+} from '@sagebionetworks/openchallenges/ui';
 import { ConfigService } from '@sagebionetworks/openchallenges/config';
 import {
   HttpStatusRedirect,
   handleHttpError,
 } from '@sagebionetworks/openchallenges/util';
+import { MatTabsModule } from '@angular/material/tabs';
+import { MatIconModule } from '@angular/material/icon';
+import { ChallengeContributorsComponent } from './challenge-contributors/challenge-contributors.component';
+import { ChallengeOrganizersComponent } from './challenge-organizers/challenge-organizers.component';
+import { ChallengeOverviewComponent } from './challenge-overview/challenge-overview.component';
+import { ChallengeStargazersComponent } from './challenge-stargazers/challenge-stargazers.component';
+import { ChallengeStatsComponent } from './challenge-stats/challenge-stats.component';
+import { CommonModule } from '@angular/common';
 
 @Component({
   selector: 'openchallenges-challenge',
+  standalone: true,
+  imports: [
+    CommonModule,
+    RouterModule,
+    MatTabsModule,
+    MatIconModule,
+    ChallengeOverviewComponent,
+    ChallengeOrganizersComponent,
+    ChallengeContributorsComponent,
+    ChallengeStargazersComponent,
+    ChallengeStatsComponent,
+    UiModule,
+    AvatarComponent,
+    FooterComponent,
+  ],
   templateUrl: './challenge.component.html',
   styleUrls: ['./challenge.component.scss'],
 })
