+++ libs/openchallenges/challenge/src/lib/challenge.routes.ts
@@ -1,0 +1,10 @@
+import { Routes } from '@angular/router';
+import { ChallengeComponent } from './challenge.component';
+
+export const routes: Routes = [
+  {
+    path: '',
+    component: ChallengeComponent,
+    children: [{ path: '**', component: ChallengeComponent }],
+  },
+];
